refactor: triton, move SD struct from config to triton discovery package, fix test
