+++ config/config.go
@@ -29,6 +29,7 @@ import (
 	"github.com/prometheus/prometheus/discovery/file"
 	"github.com/prometheus/prometheus/discovery/gce"
 	"github.com/prometheus/prometheus/discovery/marathon"
+	"github.com/prometheus/prometheus/discovery/triton"
 
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/pkg/targetgroup"
@@ -126,13 +127,6 @@ var (
 	DefaultOpenstackSDConfig = OpenstackSDConfig{
 		Port:            80,
 		RefreshInterval: model.Duration(60 * time.Second),
-	}
-
-	// DefaultTritonSDConfig is the default Triton SD configuration.
-	DefaultTritonSDConfig = TritonSDConfig{
-		Port:            9163,
-		RefreshInterval: model.Duration(60 * time.Second),
-		Version:         1,
 	}
 
 	// DefaultRemoteWriteConfig is the default remote write configuration.
@@ -381,7 +375,7 @@ type ServiceDiscoveryConfig struct {
 	// List of Azure service discovery configurations.
 	AzureSDConfigs []*azure.SDConfig `yaml:"azure_sd_configs,omitempty"`
 	// List of Triton service discovery configurations.
-	TritonSDConfigs []*TritonSDConfig `yaml:"triton_sd_configs,omitempty"`
+	TritonSDConfigs []*triton.SDConfig `yaml:"triton_sd_configs,omitempty"`
 
 	// Catches all undefined fields and must be empty after parsing.
 	XXX map[string]interface{} `yaml:",inline"`
@@ -779,42 +773,6 @@ func (c *OpenstackSDConfig) UnmarshalYAML(unmarshal func(interface{}) error) err
 		return fmt.Errorf("role missing (one of: instance, hypervisor)")
 	}
 	return yamlUtil.CheckOverflow(c.XXX, "openstack_sd_config")
-}
-
-// TritonSDConfig is the configuration for Triton based service discovery.
-type TritonSDConfig struct {
-	Account         string               `yaml:"account"`
-	DNSSuffix       string               `yaml:"dns_suffix"`
-	Endpoint        string               `yaml:"endpoint"`
-	Port            int                  `yaml:"port"`
-	RefreshInterval model.Duration       `yaml:"refresh_interval,omitempty"`
-	TLSConfig       configUtil.TLSConfig `yaml:"tls_config,omitempty"`
-	Version         int                  `yaml:"version"`
-	// Catches all undefined fields and must be empty after parsing.
-	XXX map[string]interface{} `yaml:",inline"`
-}
-
-// UnmarshalYAML implements the yaml.Unmarshaler interface.
-func (c *TritonSDConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	*c = DefaultTritonSDConfig
-	type plain TritonSDConfig
-	err := unmarshal((*plain)(c))
-	if err != nil {
-		return err
-	}
-	if c.Account == "" {
-		return fmt.Errorf("Triton SD configuration requires an account")
-	}
-	if c.DNSSuffix == "" {
-		return fmt.Errorf("Triton SD configuration requires a dns_suffix")
-	}
-	if c.Endpoint == "" {
-		return fmt.Errorf("Triton SD configuration requires an endpoint")
-	}
-	if c.RefreshInterval <= 0 {
-		return fmt.Errorf("Triton SD configuration requires RefreshInterval to be a positive integer")
-	}
-	return yamlUtil.CheckOverflow(c.XXX, "triton_sd_config")
 }
 
 // RelabelAction is the action to be performed on relabeling.
